requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
crispy-bootstrap5==0.7
Django==4.2.1
django-crispy-forms==2.0
django-recaptcha==3.0.0
django-tinymce==3.6.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
yarl==1.9.2
